Updated verbly & libtwitter++ to remove dependence on json submodule
The current version of libtwitter++ has an updated API, so the bot had
to be updated to use the new API; additionally, the usage of the
UNIX-dependent sleep call was replaced by the sleep functionality in the
C++ standard library. At some point, we should also replace the usage of
the C random library with the C++ random library. Finally, some
initialization functionality that was being repeated for every generated
tweet was instead moved out of the main loop.